more examples changes
