+++ examples/src/examples/camera/fly.mjs
@@ -1,4 +1,4 @@
-import * as pc from '../../../../';
+import * as pc from 'playcanvas';
 
 
 class FlyExample {
