+++ examples/src/examples/sound/positional.mjs
@@ -1,4 +1,4 @@
-import * as pc from '../../../../';
+import * as pc from 'playcanvas';
 
 
 class PositionalExample {
