+++ examples/src/examples/user-interface/button-basic.mjs
@@ -1,4 +1,4 @@
-import * as pc from '../../../../';
+import * as pc from 'playcanvas';
 
 class ButtonBasicExample {
     static CATEGORY = 'User Interface';
