+++ examples/src/examples/user-interface/button-sprite.mjs
@@ -1,4 +1,4 @@
-import * as pc from '../../../../';
+import * as pc from 'playcanvas';
 
 class ButtonSpriteExample {
     static CATEGORY = 'User Interface';
