+++ examples/src/examples/user-interface/custom-shader.mjs
@@ -1,4 +1,4 @@
-import * as pc from '../../../../';
+import * as pc from 'playcanvas';
 
 class CustomShaderExample {
     static CATEGORY = 'User Interface';
