+++ examples/src/examples/user-interface/text-emojis.mjs
@@ -1,14 +1,13 @@
-import * as pc from '../../../../';
+import * as pc from 'playcanvas';
 
-class LayoutGroupExample {
+class TextEmojisExample {
     static CATEGORY = 'User Interface';
-    static NAME = 'Layout Group';
-    static WEBGPU_ENABLED = true;
+    static NAME = 'Text Emojis';
 
     example(canvas: HTMLCanvasElement, deviceType: string): void {
 
         const assets = {
-            'font': new pc.Asset('font', 'font', { url: '/static/assets/fonts/courier.json' })
+            'font': new pc.Asset('font', 'font', { url: '/static/assets/fonts/arial.json' })
         };
 
         const gfxOptions = {
@@ -38,6 +37,10 @@ class LayoutGroupExample {
                 pc.ElementComponentSystem,
                 // @ts-ignore
                 pc.LayoutGroupComponentSystem,
+                // @ts-ignore
+                pc.ScrollViewComponentSystem,
+                // @ts-ignore
+                pc.ScrollbarComponentSystem,
                 // @ts-ignore
                 pc.LayoutChildComponentSystem
             ];
@@ -81,20 +84,74 @@ class LayoutGroupExample {
                 });
                 app.root.addChild(screen);
 
+                // some sample text
+                const firstLineText = "PlayCanvas supports Emojis via CanvasFont!";
+                const flagsText = "Flags: 🇺🇸🇩🇪🇮🇪🇮🇹🏴‍☠️🇨🇦";
+                const complexText = "Complex emoji: 👨🏿3️⃣👁️‍🗨️";
+
+                // Create a canvas font asset
+                const size = 64;
+                const elSize = 32;
+
+                const canvasFont = new pc.CanvasFont(app, {
+                    color: new pc.Color(1, 1, 1), // white
+                    fontName: "Arial",
+                    fontSize: size,
+                    width: 256,
+                    height: 256
+                });
+
+                // The first texture update needs to be `createTextures()`. Follow-up calls need to be `updateTextures()`.
+                canvasFont.createTextures(firstLineText);
+                canvasFont.updateTextures(flagsText);
+                canvasFont.updateTextures(complexText);
+
+                // Create the text entities
+                function createText(y: number, text: string) {
+                    const canvasElementEntity = new pc.Entity();
+                    canvasElementEntity.setLocalPosition(0, y, 0);
+                    canvasElementEntity.addComponent("element", {
+                        pivot: new pc.Vec2(0.5, 0.5),
+                        anchor: new pc.Vec4(0.5, 0.5, 0.5, 0.5),
+                        fontSize: elSize,
+                        text: text,
+                        type: pc.ELEMENTTYPE_TEXT
+                    });
+                    canvasElementEntity.element.font = canvasFont;
+                    screen.addChild(canvasElementEntity);
+                }
+                createText(225, firstLineText);
+                createText(150, flagsText);
+                createText(100, complexText);
+
+
+                // Canvas Fonts Debug - you shouldn't do this in your actual project
+                const debugText = new pc.Entity();
+                debugText.setLocalPosition(0, -50, 0);
+                debugText.addComponent("element", {
+                    pivot: new pc.Vec2(0.5, 0.5),
+                    anchor: new pc.Vec4(0.5, 0.5, 0.5, 0.5),
+                    fontAsset: assets.font.id,
+                    fontSize: elSize,
+                    text: "The following are the CanvasFont's Texture Atlases,\ncontaining all the rendered characters:",
+                    type: pc.ELEMENTTYPE_TEXT
+                });
+                screen.addChild(debugText);
+
                 // Create Layout Group Entity
                 const group = new pc.Entity();
+                group.setLocalPosition(0, -150, 0);
                 group.addComponent("element", {
                     // a Layout Group needs a 'group' element component
                     type: pc.ELEMENTTYPE_GROUP,
                     anchor: [0.5, 0.5, 0.5, 0.5],
                     pivot: [0.5, 0.5],
                     // the element's width and height dictate the group's bounds
-                    width: 350,
-                    height: 150
+                    width: 300,
+                    height: 100
                 });
                 group.addComponent("layoutgroup", {
                     orientation: pc.ORIENTATION_HORIZONTAL,
-                    spacing: new pc.Vec2(10, 10),
                     // fit_both for width and height, making all child elements take the entire space
                     widthFitting: pc.FITTING_BOTH,
                     heightFitting: pc.FITTING_BOTH,
@@ -103,44 +160,26 @@ class LayoutGroupExample {
                 });
                 screen.addChild(group);
 
-                // create 15 children to show off the layout group
-                for (let i = 0; i < 15; ++i) {
+                // create 1 child per texture
+                for (let i = 0; i < canvasFont.textures.length; i++) {
+                    const texture = canvasFont.textures[i];
+
                     // create a random-colored panel
                     const child = new pc.Entity();
                     child.addComponent("element", {
                         anchor: [0.5, 0.5, 0.5, 0.5],
                         pivot: [0.5, 0.5],
-                        color: new pc.Color(Math.random(), Math.random(), Math.random()),
+                        texture: texture,
                         type: pc.ELEMENTTYPE_IMAGE
                     });
                     child.addComponent("layoutchild", {
                         excludeFromLayout: false
                     });
                     group.addChild(child);
-
-                    // add a text label
-                    const childLabel = new pc.Entity();
-                    childLabel.addComponent("element", {
-                        // center-position and attach to the borders of parent
-                        // meaning this text element will scale along with parent
-                        anchor: [0, 0, 1, 1],
-                        margin: [0, 0, 0, 0],
-                        pivot: [0.5, 0.5],
-                        color: new pc.Color(1, 1, 1),
-                        fontAsset: assets.font.id,
-                        text: `${i + 1}`,
-                        type: pc.ELEMENTTYPE_TEXT,
-                        // auto font size
-                        autoWidth: false,
-                        autoHeight: false,
-                        autoFitWidth: true,
-                        autoFitHeight: true
-                    });
-                    child.addChild(childLabel);
                 }
             });
         });
     }
 }
 
-export default LayoutGroupExample;
+export default TextEmojisExample;
