+++ examples/src/examples/user-interface/text-typewriter.mjs
@@ -1,4 +1,4 @@
-import * as pc from '../../../../';
+import * as pc from 'playcanvas';
 
 class TextTypewriterExample {
     static CATEGORY = 'User Interface';
