+++ examples/src/examples/user-interface/world-to-screen.mjs
@@ -1,4 +1,4 @@
-import * as pc from '../../../../';
+import * as pc from 'playcanvas';
 
 class WorldToScreenExample {
     static CATEGORY = 'User Interface';
