+++ examples/src/examples/user-interface/world-ui.mjs
@@ -1,4 +1,4 @@
-import * as pc from '../../../../';
+import * as pc from 'playcanvas';
 
 class WorldUiExample {
     static CATEGORY = 'User Interface';
