+++ examples/src/examples/xr/vr-hands.mjs
@@ -1,4 +1,4 @@
-import * as pc from '../../../../';
+import * as pc from 'playcanvas';
 
 class VrHandsExample {
     static CATEGORY = 'XR';
