+++ examples/src/examples/xr/vr-movement.mjs
@@ -1,4 +1,4 @@
-import * as pc from '../../../../';
+import * as pc from 'playcanvas';
 
 class VrMovementExample {
     static CATEGORY = 'XR';
