+++ examples/src/examples/xr/xr-picking.mjs
@@ -1,4 +1,4 @@
-import * as pc from '../../../../';
+import * as pc from 'playcanvas';
 
 class XrPickingExample {
     static CATEGORY = 'XR';
